refactor(analytics): Move table/slice definitions to tables file
